+++ skactiveml/classifier/multiannotator/__init__.py
@@ -1,5 +1,5 @@
 from ._annotator_ensemble_classifier import AnnotatorEnsembleClassifier
 from ._annotator_logistic_regression import AnnotatorLogisticRegression
-from ._crowd_layer import CrowdLayerClassifier
+from ._crowd_layer_classifier import CrowdLayerClassifier
 
 __all__ = ["AnnotatorLogisticRegression", "AnnotatorEnsembleClassifier", "CrowdLayerClassifier"]
